+++ src/data/duke.txt
@@ -1,3 +1,2 @@
-Todo: run     (｡-_-｡ ) INCOMPLETE ( ｡-_-｡)    git 
-Event: project meeting from 01/01/2023 1200 to 01/01/2023 1300     (｡-_-｡ ) INCOMPLETE ( ｡-_-｡)    git 
-Deadline: return book by 02/01/2023 1230     (｡-_-｡ ) INCOMPLETE ( ｡-_-｡)    git 
+Todo: run     ๑(◕‿◕)๑ COMPLETED! ๑(◕‿◕)๑    
+Deadline: interview internship by 01/01/2023 1300     (｡-_-｡ ) INCOMPLETE ( ｡-_-｡)    git
