+++ src/main/java/duke/task/Todo.java
@@ -1,10 +1,10 @@
-/**
- * File name: Todo.java
- * @author: Jerome Neo
- * Description: Todo class that inherits from the Task class.
- */
 package duke.task;
 
+/**
+ *      File name: Todo.java
+ *      @author: Jerome Neo
+ *      Description: Todo class that inherits from the Task class.
+ */
 public class Todo extends Task {
 
     /**
@@ -23,7 +23,7 @@ public class Todo extends Task {
      */
     @Override
     public String toString() {
-        return "[T]" + "[" +this.getStatusIcon() + "] " + this.getDescription();
+        return "[T]" + "[" + this.getStatusIcon() + "] " + this.getDescription();
     }
 }
